+++ commonjs/index.js
@@ -1,7 +1,7 @@
 "use strict";
 // TODO: Switch all the constants to actual enums.
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.WebBrowserFacetLinkType = exports.VanillaGameplayUIProfile = exports.VanillaGameplayContainerItemType = exports.VanillaGameplayContainerChestType = exports.RealmsPermissionRole = exports.RealmsPermissionAction = exports.PlayerTitleHistory = exports.PlayerPartyPresence = exports.PlayerRelation = exports.PlayerPermissionLevel = exports.RealmsStoriesTimelineOptInStatus = exports.FriendFavoriteStatus = exports.FriendPresence = exports.RealmsPlayerListPlayerPermission = exports.RealmPlayerRoleEnum = exports.StorageType = exports.KeyboardType = exports.Controller = exports.WorldPlayerInfoBindingsConnectionType = exports.WorldPlayerInfoBindingsPlatform = exports.Platform = exports.ARVRPlatform = exports.InputMethod = exports.HandheldDeviceType = exports.ScreenType = exports.LeaveRealmsServerError = exports.JoinRealmsServerError = void 0;
+exports.DownloadWorldTemplateError = exports.ImportFailure = exports.DownloadWorldTemplateStatus = exports.WorldSizeConvertResult = exports.StartClearPlayerDataError = exports.ExportWorldStatus = exports.ClearPlayerDataType = exports.ExportWorldResult = exports.ExportWorldFlags = exports.DuplicateWorldError = exports.FacetTaskState = exports.InvocationResult = exports.InvocationState = exports.WorldCloudSyncResult = exports.FriendsLoadingState = exports.ProfileImageState = exports.WorldPingStatus = exports.NetworkWorldType = exports.DeletionStatus = exports.PostStatus = exports.FetchStatus = exports.WebBrowserLink = exports.VanillaGameplayUIProfile = exports.VanillaGameplayContainerItemType = exports.VanillaGameplayContainerChestType = exports.RealmsPermissionRole = exports.RealmsPermissionAction = exports.PlayerTitleHistory = exports.PartyPresence = exports.PlayerRelation = exports.PlayerPermissionLevel = exports.RealmsStoriesTimelineOptInStatus = exports.FavoriteStatusEnum = exports.SocialPresence = exports.RealmsPlayerListPlayerPermission = exports.RealmPlayerRoleEnum = exports.StorageType = exports.KeyboardType = exports.Controller = exports.WorldPlayerInfoBindingsConnectionType = exports.WorldPlayerInfoBindingsPlatform = exports.Platform = exports.ARVRPlatform = exports.InputMethod = exports.HandheldDeviceType = exports.ScreenType = exports.LeaveRealmsServerError = exports.JoinRealmsServerError = void 0;
 exports.reverseNumericEnumMapping = reverseNumericEnumMapping;
 exports.JoinRealmsServerError = {
     RealmsAPIUnavailable: 0,
@@ -123,13 +123,13 @@ exports.RealmsPlayerListPlayerPermission = {
     OPERATOR: 3,
     OWNER: 4,
 };
-exports.FriendPresence = {
+exports.SocialPresence = {
     Unknown: 0,
     Online: 1,
     Away: 2,
     Offline: 3,
 };
-exports.FriendFavoriteStatus = {
+exports.FavoriteStatusEnum = {
     UNKNOWN: 0,
     FAVORITE: 1,
     NOT_FAVORITE: 2,
@@ -166,7 +166,7 @@ exports.PlayerRelation = {
 //     NumberEnumToObject<typeof PlayerRelation>,
 //     Mode
 // >;
-exports.PlayerPartyPresence = {
+exports.PartyPresence = {
     NotInParty: 0,
     InParty: 1,
     Disconnecting: 2,
@@ -221,7 +221,7 @@ exports.VanillaGameplayUIProfile = {
     Pocket: 1,
     None: 2,
 };
-exports.WebBrowserFacetLinkType = {
+exports.WebBrowserLink = {
     NewWorldCreationFeedbackPage: 0,
     PlayScreenFeedbackPage: 1,
     EditWorldScreenFeedbackPage: 2,
@@ -258,6 +258,162 @@ exports.WebBrowserFacetLinkType = {
     AndroidGooglePlay: 33,
     AndroidAmazonAppStore: 34,
     XboxOneStore: 35,
+};
+exports.FetchStatus = {
+    Idle: 0,
+    InProgress: 1,
+    Success: 2,
+    InternalError: 3,
+    FetchFailed: 4,
+    RateLimit: 5,
+};
+exports.PostStatus = {
+    Idle: 0,
+    InProgress: 1,
+    Success: 2,
+    InternalError: 3,
+    InputInvalid: 4,
+    Uploading: 5,
+    UploadFailed: 6,
+    PostingFailed: 7,
+    RateLimit: 8,
+};
+exports.DeletionStatus = {
+    Idle: 0,
+    DeleteComment: 1,
+    DeletePost: 2,
+    Dismiss: 3,
+    isDeleting: 4,
+    isDismissing: 5,
+};
+exports.NetworkWorldType = {
+    ThirdParty: 0,
+    External: 1,
+    Realm: 2,
+    LAN: 3,
+};
+exports.WorldPingStatus = {
+    Unavailable: 0,
+    Low: 1,
+    Medium: 2,
+    High: 3,
+};
+exports.ProfileImageState = {
+    Undefined: 0,
+    Requested: 1,
+    Success: 2,
+    Failure: 3,
+};
+exports.FriendsLoadingState = {
+    Unavailable: 0,
+    Loading: 1,
+    Ready: 2,
+    Error: 3,
+};
+exports.WorldCloudSyncResult = {
+    Ok: 0,
+    Unsupported: 1,
+    WorldNotCloudStored: 2,
+    UnableToGetManifest: 3,
+    UnableToSyncWorld: 4,
+};
+/**
+ * Represents the state of a command invocation.
+ */
+exports.InvocationState = {
+    Idle: 0,
+    Working: 1,
+    Done: 2,
+};
+exports.InvocationResult = {
+    Invalid: -1,
+    Success: 0,
+    Fail: 1,
+};
+/**
+ * Represents the state of a facet method execution.
+ *
+ * A facet that has a method that should be possible to sequence using
+ * `useFacetCoroutine` should expose a corresponding state property of this
+ * type that is updated by the facet when the method is running.
+ *
+ * example:
+ * ```
+ * type NetworkFacet = {
+ *  uploadToServer(): void
+ *  uploadToServerProgress: FacetTaskState
+ * }
+ * ```
+ */
+exports.FacetTaskState = {
+    IDLE: 0,
+    RUNNING: 1,
+    DONE: 2,
+    CANCELLED: 3,
+    FAILED: 4,
+};
+exports.DuplicateWorldError = {
+    Ok: 0,
+    IncorrectWorldId: 1,
+    InsufficientDiskSpace: 2,
+    UnknownError: 3,
+};
+exports.ExportWorldFlags = {
+    None: 0,
+    ClearPlayerData: 2,
+    ForceHardcoreMode: 4,
+    All: -1,
+};
+exports.ExportWorldResult = {
+    IncorrectWorldId: 1,
+    NoFile: 2,
+    ZipError: 3,
+    PremiumContent: 4,
+    EditionMismatch: 5,
+    EditorMismatch: 6,
+    ZipDepthError: 7,
+    UnknownError: 8,
+    Cancelled: 9,
+};
+exports.ClearPlayerDataType = {
+    SkipLocalUser: 0,
+    IncludeLocalUser: 1,
+};
+exports.ExportWorldStatus = {
+    Idle: 0,
+    Setup: 1,
+    SelectingFileDestination: 2,
+    Exporting: 3,
+};
+exports.StartClearPlayerDataError = {
+    FailedToOpenDatabase: 0,
+};
+exports.WorldSizeConvertResult = {
+    Ok: 0,
+    InvalidLevel: 1,
+};
+exports.DownloadWorldTemplateStatus = {
+    Initialiing: 0,
+    Downloading: 1,
+    Importing: 2,
+    Success: 3,
+    Fail: 4,
+};
+exports.ImportFailure = {
+    UnknownFileType: 0,
+    Duplicate: 1,
+    MalformedZip: 2,
+    Default: 3,
+    Incompatible: 4,
+    Cancelled: 5,
+};
+exports.DownloadWorldTemplateError = {
+    AlreadyDownloaded: 0,
+    DownloadAlreadyInProgress: 1,
+    DownloadFailed: 2,
+    ItemNotOwned: 3,
+    CancelledByUser: 4,
+    InsufficientStorage: 5,
 };
 /**
  * Reverses a numeric enum mapping.