+++ index.d.ts
@@ -168,19 +168,19 @@ export declare const RealmsPlayerListPlayerPermission: {
     readonly OWNER: 4;
 };
 export type RealmsPlayerListPlayerPermission<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof RealmsPlayerListPlayerPermission, Mode>;
-export declare const FriendPresence: {
+export declare const SocialPresence: {
     readonly Unknown: 0;
     readonly Online: 1;
     readonly Away: 2;
     readonly Offline: 3;
 };
-export type FriendPresence<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FriendPresence, Mode>;
-export declare const FriendFavoriteStatus: {
+export type SocialPresence<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof SocialPresence, Mode>;
+export declare const FavoriteStatusEnum: {
     readonly UNKNOWN: 0;
     readonly FAVORITE: 1;
     readonly NOT_FAVORITE: 2;
 };
-export type FriendFavoriteStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FriendFavoriteStatus, Mode>;
+export type FavoriteStatusEnum<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FavoriteStatusEnum, Mode>;
 export declare const RealmsStoriesTimelineOptInStatus: {
     readonly OptedIn: 0;
     readonly OptedOut: 1;
@@ -203,12 +203,12 @@ export declare const PlayerRelation: {
     readonly Unknown: 5;
 };
 export type PlayerRelation<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof PlayerRelation, Mode>;
-export declare const PlayerPartyPresence: {
+export declare const PartyPresence: {
     readonly NotInParty: 0;
     readonly InParty: 1;
     readonly Disconnecting: 2;
 };
-export type PlayerPartyPresence<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof PlayerPartyPresence, Mode>;
+export type PartyPresence<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof PartyPresence, Mode>;
 export declare const PlayerTitleHistory: {
     readonly Unknown: 0;
     readonly NotPlayed: 1;
@@ -265,7 +265,7 @@ export declare const VanillaGameplayUIProfile: {
     readonly None: 2;
 };
 export type VanillaGameplayUIProfile<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof VanillaGameplayUIProfile, Mode>;
-export declare const WebBrowserFacetLinkType: {
+export declare const WebBrowserLink: {
     readonly NewWorldCreationFeedbackPage: 0;
     readonly PlayScreenFeedbackPage: 1;
     readonly EditWorldScreenFeedbackPage: 2;
@@ -303,7 +303,184 @@ export declare const WebBrowserFacetLinkType: {
     readonly AndroidAmazonAppStore: 34;
     readonly XboxOneStore: 35;
 };
-export type WebBrowserFacetLinkType<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof WebBrowserFacetLinkType, Mode>;
+export type WebBrowserLink<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof WebBrowserLink, Mode>;
+export declare const FetchStatus: {
+    readonly Idle: 0;
+    readonly InProgress: 1;
+    readonly Success: 2;
+    readonly InternalError: 3;
+    readonly FetchFailed: 4;
+    readonly RateLimit: 5;
+};
+export type FetchStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FetchStatus, Mode>;
+export declare const PostStatus: {
+    readonly Idle: 0;
+    readonly InProgress: 1;
+    readonly Success: 2;
+    readonly InternalError: 3;
+    readonly InputInvalid: 4;
+    readonly Uploading: 5;
+    readonly UploadFailed: 6;
+    readonly PostingFailed: 7;
+    readonly RateLimit: 8;
+};
+export type PostStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof PostStatus, Mode>;
+export declare const DeletionStatus: {
+    readonly Idle: 0;
+    readonly DeleteComment: 1;
+    readonly DeletePost: 2;
+    readonly Dismiss: 3;
+    readonly isDeleting: 4;
+    readonly isDismissing: 5;
+};
+export type DeletionStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof DeletionStatus, Mode>;
+export declare const NetworkWorldType: {
+    readonly ThirdParty: 0;
+    readonly External: 1;
+    readonly Realm: 2;
+    readonly LAN: 3;
+};
+export type NetworkWorldType<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof NetworkWorldType, Mode>;
+export declare const WorldPingStatus: {
+    readonly Unavailable: 0;
+    readonly Low: 1;
+    readonly Medium: 2;
+    readonly High: 3;
+};
+export type WorldPingStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof WorldPingStatus, Mode>;
+export declare const ProfileImageState: {
+    readonly Undefined: 0;
+    readonly Requested: 1;
+    readonly Success: 2;
+    readonly Failure: 3;
+};
+export type ProfileImageState<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ProfileImageState, Mode>;
+export declare const FriendsLoadingState: {
+    readonly Unavailable: 0;
+    readonly Loading: 1;
+    readonly Ready: 2;
+    readonly Error: 3;
+};
+export type FriendsLoadingState<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FriendsLoadingState, Mode>;
+export declare const WorldCloudSyncResult: {
+    readonly Ok: 0;
+    readonly Unsupported: 1;
+    readonly WorldNotCloudStored: 2;
+    readonly UnableToGetManifest: 3;
+    readonly UnableToSyncWorld: 4;
+};
+export type WorldCloudSyncResult<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof WorldCloudSyncResult, Mode>;
+/**
+ * Represents the state of a command invocation.
+ */
+export declare const InvocationState: {
+    readonly Idle: 0;
+    readonly Working: 1;
+    readonly Done: 2;
+};
+export type InvocationState<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof InvocationState, Mode>;
+export declare const InvocationResult: {
+    readonly Invalid: -1;
+    readonly Success: 0;
+    readonly Fail: 1;
+};
+export type InvocationResult<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof InvocationResult, Mode>;
+/**
+ * Represents the state of a facet method execution.
+ *
+ * A facet that has a method that should be possible to sequence using
+ * `useFacetCoroutine` should expose a corresponding state property of this
+ * type that is updated by the facet when the method is running.
+ *
+ * example:
+ * ```
+ * type NetworkFacet = {
+ *  uploadToServer(): void
+ *  uploadToServerProgress: FacetTaskState
+ * }
+ * ```
+ */
+export declare const FacetTaskState: {
+    readonly IDLE: 0;
+    readonly RUNNING: 1;
+    readonly DONE: 2;
+    readonly CANCELLED: 3;
+    readonly FAILED: 4;
+};
+export type FacetTaskState<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FacetTaskState, Mode>;
+export declare const DuplicateWorldError: {
+    readonly Ok: 0;
+    readonly IncorrectWorldId: 1;
+    readonly InsufficientDiskSpace: 2;
+    readonly UnknownError: 3;
+};
+export type DuplicateWorldError<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof DuplicateWorldError, Mode>;
+export declare const ExportWorldFlags: {
+    readonly None: 0;
+    readonly ClearPlayerData: 2;
+    readonly ForceHardcoreMode: 4;
+    readonly All: -1;
+};
+export type ExportWorldFlags<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ExportWorldFlags, Mode>;
+export declare const ExportWorldResult: {
+    readonly IncorrectWorldId: 1;
+    readonly NoFile: 2;
+    readonly ZipError: 3;
+    readonly PremiumContent: 4;
+    readonly EditionMismatch: 5;
+    readonly EditorMismatch: 6;
+    readonly ZipDepthError: 7;
+    readonly UnknownError: 8;
+    readonly Cancelled: 9;
+};
+export type ExportWorldResult<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ExportWorldResult, Mode>;
+export declare const ClearPlayerDataType: {
+    readonly SkipLocalUser: 0;
+    readonly IncludeLocalUser: 1;
+};
+export type ClearPlayerDataType<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ClearPlayerDataType, Mode>;
+export declare const ExportWorldStatus: {
+    readonly Idle: 0;
+    readonly Setup: 1;
+    readonly SelectingFileDestination: 2;
+    readonly Exporting: 3;
+};
+export type ExportWorldStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ExportWorldStatus, Mode>;
+export declare const StartClearPlayerDataError: {
+    readonly FailedToOpenDatabase: 0;
+};
+export type StartClearPlayerDataError<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof StartClearPlayerDataError, Mode>;
+export declare const WorldSizeConvertResult: {
+    readonly Ok: 0;
+    readonly InvalidLevel: 1;
+};
+export type WorldSizeConvertResult<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof WorldSizeConvertResult, Mode>;
+export declare const DownloadWorldTemplateStatus: {
+    readonly Initialiing: 0;
+    readonly Downloading: 1;
+    readonly Importing: 2;
+    readonly Success: 3;
+    readonly Fail: 4;
+};
+export type DownloadWorldTemplateStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof DownloadWorldTemplateStatus, Mode>;
+export declare const ImportFailure: {
+    readonly UnknownFileType: 0;
+    readonly Duplicate: 1;
+    readonly MalformedZip: 2;
+    readonly Default: 3;
+    readonly Incompatible: 4;
+    readonly Cancelled: 5;
+};
+export type ImportFailure<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ImportFailure, Mode>;
+export declare const DownloadWorldTemplateError: {
+    readonly AlreadyDownloaded: 0;
+    readonly DownloadAlreadyInProgress: 1;
+    readonly DownloadFailed: 2;
+    readonly ItemNotOwned: 3;
+    readonly CancelledByUser: 4;
+    readonly InsufficientStorage: 5;
+};
+export type DownloadWorldTemplateError<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof DownloadWorldTemplateError, Mode>;
 /**
  * Reverses a numeric enum mapping.
  *