+++ index.ts
@@ -240,20 +240,20 @@ export type RealmsPlayerListPlayerPermission<Mode extends "enum" | "keys" | "val
     Mode
 >;
 
-export const FriendPresence = {
+export const SocialPresence = {
     Unknown: 0,
     Online: 1,
     Away: 2,
     Offline: 3,
 } as const;
-export type FriendPresence<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FriendPresence, Mode>;
+export type SocialPresence<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof SocialPresence, Mode>;
 
-export const FriendFavoriteStatus = {
+export const FavoriteStatusEnum = {
     UNKNOWN: 0,
     FAVORITE: 1,
     NOT_FAVORITE: 2,
 } as const;
-export type FriendFavoriteStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FriendFavoriteStatus, Mode>;
+export type FavoriteStatusEnum<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FavoriteStatusEnum, Mode>;
 
 export const RealmsStoriesTimelineOptInStatus = {
     OptedIn: 0,
@@ -300,12 +300,12 @@ export type PlayerRelation<Mode extends "enum" | "keys" | "values" = "enum"> = C
 //     Mode
 // >;
 
-export const PlayerPartyPresence = {
+export const PartyPresence = {
     NotInParty: 0,
     InParty: 1,
     Disconnecting: 2,
 } as const;
-export type PlayerPartyPresence<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof PlayerPartyPresence, Mode>;
+export type PartyPresence<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof PartyPresence, Mode>;
 
 export const PlayerTitleHistory = {
     Unknown: 0,
@@ -381,7 +381,7 @@ export type VanillaGameplayUIProfile<Mode extends "enum" | "keys" | "values" = "
     Mode
 >;
 
-export const WebBrowserFacetLinkType = {
+export const WebBrowserLink = {
     NewWorldCreationFeedbackPage: 0,
     PlayScreenFeedbackPage: 1,
     EditWorldScreenFeedbackPage: 2,
@@ -419,8 +419,215 @@ export const WebBrowserFacetLinkType = {
     AndroidAmazonAppStore: 34,
     XboxOneStore: 35,
 } as const;
-export type WebBrowserFacetLinkType<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<
-    typeof WebBrowserFacetLinkType,
+export type WebBrowserLink<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof WebBrowserLink, Mode>;
+
+export const FetchStatus = {
+    Idle: 0,
+    InProgress: 1,
+    Success: 2,
+    InternalError: 3,
+    FetchFailed: 4,
+    RateLimit: 5,
+} as const;
+export type FetchStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FetchStatus, Mode>;
+
+export const PostStatus = {
+    Idle: 0,
+    InProgress: 1,
+    Success: 2,
+    InternalError: 3,
+    InputInvalid: 4,
+    Uploading: 5,
+    UploadFailed: 6,
+    PostingFailed: 7,
+    RateLimit: 8,
+} as const;
+export type PostStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof PostStatus, Mode>;
+
+export const DeletionStatus = {
+    Idle: 0,
+    DeleteComment: 1,
+    DeletePost: 2,
+    Dismiss: 3,
+    isDeleting: 4,
+    isDismissing: 5,
+} as const;
+export type DeletionStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof DeletionStatus, Mode>;
+
+export const NetworkWorldType = {
+    ThirdParty: 0,
+    External: 1,
+    Realm: 2,
+    LAN: 3,
+} as const;
+export type NetworkWorldType<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof NetworkWorldType, Mode>;
+
+export const WorldPingStatus = {
+    Unavailable: 0,
+    Low: 1,
+    Medium: 2,
+    High: 3,
+} as const;
+export type WorldPingStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof WorldPingStatus, Mode>;
+
+export const ProfileImageState = {
+    Undefined: 0,
+    Requested: 1,
+    Success: 2,
+    Failure: 3,
+} as const;
+export type ProfileImageState<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ProfileImageState, Mode>;
+
+export const FriendsLoadingState = {
+    Unavailable: 0,
+    Loading: 1,
+    Ready: 2,
+    Error: 3,
+} as const;
+export type FriendsLoadingState<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FriendsLoadingState, Mode>;
+
+export const WorldCloudSyncResult = {
+    Ok: 0,
+    Unsupported: 1,
+    WorldNotCloudStored: 2,
+    UnableToGetManifest: 3,
+    UnableToSyncWorld: 4,
+} as const;
+export type WorldCloudSyncResult<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof WorldCloudSyncResult, Mode>;
+
+/**
+ * Represents the state of a command invocation.
+ */
+export const InvocationState = {
+    Idle: 0,
+    Working: 1,
+    Done: 2,
+} as const;
+export type InvocationState<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof InvocationState, Mode>;
+
+export const InvocationResult = {
+    Invalid: -1,
+    Success: 0,
+    Fail: 1,
+} as const;
+export type InvocationResult<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof InvocationResult, Mode>;
+
+/**
+ * Represents the state of a facet method execution.
+ *
+ * A facet that has a method that should be possible to sequence using
+ * `useFacetCoroutine` should expose a corresponding state property of this
+ * type that is updated by the facet when the method is running.
+ *
+ * example:
+ * ```
+ * type NetworkFacet = {
+ *  uploadToServer(): void
+ *  uploadToServerProgress: FacetTaskState
+ * }
+ * ```
+ */
+export const FacetTaskState = {
+    IDLE: 0,
+    RUNNING: 1,
+    DONE: 2,
+    CANCELLED: 3,
+    FAILED: 4,
+} as const;
+export type FacetTaskState<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof FacetTaskState, Mode>;
+
+export const DuplicateWorldError = {
+    Ok: 0,
+    IncorrectWorldId: 1,
+    InsufficientDiskSpace: 2,
+    UnknownError: 3,
+} as const;
+export type DuplicateWorldError<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof DuplicateWorldError, Mode>;
+
+export const ExportWorldFlags = {
+    None: 0,
+    ClearPlayerData: 2,
+    ForceHardcoreMode: 4,
+    All: -1,
+} as const;
+export type ExportWorldFlags<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ExportWorldFlags, Mode>;
+
+export const ExportWorldResult = {
+    IncorrectWorldId: 1,
+    NoFile: 2,
+    ZipError: 3,
+    PremiumContent: 4,
+    EditionMismatch: 5,
+    EditorMismatch: 6,
+    ZipDepthError: 7,
+    UnknownError: 8,
+    Cancelled: 9,
+} as const;
+export type ExportWorldResult<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ExportWorldResult, Mode>;
+
+export const ClearPlayerDataType = {
+    SkipLocalUser: 0,
+    IncludeLocalUser: 1,
+} as const;
+export type ClearPlayerDataType<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ClearPlayerDataType, Mode>;
+
+export const ExportWorldStatus = {
+    Idle: 0,
+    Setup: 1,
+    SelectingFileDestination: 2,
+    Exporting: 3,
+} as const;
+export type ExportWorldStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ExportWorldStatus, Mode>;
+
+export const StartClearPlayerDataError = {
+    FailedToOpenDatabase: 0,
+} as const;
+export type StartClearPlayerDataError<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<
+    typeof StartClearPlayerDataError,
+    Mode
+>;
+
+export const WorldSizeConvertResult = {
+    Ok: 0,
+    InvalidLevel: 1,
+} as const;
+export type WorldSizeConvertResult<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<
+    typeof WorldSizeConvertResult,
+    Mode
+>;
+
+export const DownloadWorldTemplateStatus = {
+    Initialiing: 0,
+    Downloading: 1,
+    Importing: 2,
+    Success: 3,
+    Fail: 4,
+} as const;
+export type DownloadWorldTemplateStatus<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<
+    typeof DownloadWorldTemplateStatus,
+    Mode
+>;
+
+export const ImportFailure = {
+    UnknownFileType: 0,
+    Duplicate: 1,
+    MalformedZip: 2,
+    Default: 3,
+    Incompatible: 4,
+    Cancelled: 5,
+} as const;
+export type ImportFailure<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<typeof ImportFailure, Mode>;
+
+export const DownloadWorldTemplateError = {
+    AlreadyDownloaded: 0,
+    DownloadAlreadyInProgress: 1,
+    DownloadFailed: 2,
+    ItemNotOwned: 3,
+    CancelledByUser: 4,
+    InsufficientStorage: 5,
+} as const;
+export type DownloadWorldTemplateError<Mode extends "enum" | "keys" | "values" = "enum"> = ConstNumberObjectEnumToEnumMappingType<
+    typeof DownloadWorldTemplateError,
     Mode
 >;