+++ Changelog.md
@@ -1,3 +1,37 @@
+# v1.7.0
+
+## Breaking Changes
+
+-   Renamed the `FriendFavoriteStatus` enum constant to `FavoriteStatusEnum`.
+-   Renamed the `FriendPresence` enum constant to `SocialPresence`.
+-   Renamed the `PlayerPartyPresence` enum constant to `PartyPresence`.
+-   Renamed the `WebBrowserFacetLinkType` enum constant to `WebBrowserLink`.
+
+## Additions
+
+-   Added the following enum constants:
+    -   `FetchStatus`
+    -   `PostStatus`
+    -   `DeletionStatus`
+    -   `NetworkWorldType`
+    -   `WorldPingStatus`
+    -   `ProfileImageState`
+    -   `FriendsLoadingState`
+    -   `WorldCloudSyncResult`
+    -   `InvocationState`
+    -   `InvocationResult`
+    -   `FacetTaskState`
+    -   `DuplicateWorldError`
+    -   `ExportWorldFlags`
+    -   `ExportWorldResult`
+    -   `ClearPlayerDataType`
+    -   `ExportWorldStatus`
+    -   `StartClearPlayerDataError`
+    -   `WorldSizeConvertResult`
+    -   `DownloadWorldTemplateStatus`
+    -   `ImportFailure`
+    -   `DownloadWorldTemplateError`
+
 # v1.6.0
 
 ## Additions